scheduler: DesiredCanaries can be set on every pass safely
The reconcile loop sets `DeploymentState.DesiredCanaries` only on the first
pass through the loop. In MRD, deployments will make one pass though the loop
while "pending", and were not ever getting `DesiredCanaries` set. This value
is static for a given version of a job because it's coming from the update
stanza, so it's entirely safe to re-assign the value on subsequent passes.